Move version to its own file and improve created_by_version explanation
