+++ pg_search/src/postgres/storage/metadata.rs
@@ -15,6 +15,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
+use crate::api::version::{parse_version_component, Version};
 use crate::postgres::rel::PgSearchRelation;
 use crate::postgres::storage::block::{block_number_is_valid, SegmentMetaEntry};
 use crate::postgres::storage::buffer::{
@@ -28,46 +29,6 @@ use pgrx::{
     function_name, iter::TableIterator, name, pg_extern, pg_sys, PgLogLevel, PgRelation,
     PgSqlErrorCode,
 };
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-#[allow(dead_code)]
-pub struct Version {
-    pub major: u16,
-    pub minor: u16,
-    pub patch: u16,
-}
-
-impl Version {
-    #[allow(dead_code)]
-    pub fn new(major: u16, minor: u16, patch: u16) -> Self {
-        Self {
-            major,
-            minor,
-            patch,
-        }
-    }
-}
-
-impl std::fmt::Display for Version {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}.{}.{}", self.major, self.minor, self.patch)
-    }
-}
-
-/// Parses a decimal version component at compile time. Overflow or non-digit bytes
-/// produce a compile error via the const-eval rules.
-const fn parse_version_component(s: &str) -> u16 {
-    let bytes = s.as_bytes();
-    let mut result: u16 = 0;
-    let mut i = 0;
-    while i < bytes.len() {
-        let b = bytes[i];
-        assert!(b >= b'0' && b <= b'9', "version component must be decimal");
-        result = result * 10 + (b - b'0') as u16;
-        i += 1;
-    }
-    result
-}
 
 /// The metadata stored on the [`Metadata`] page
 #[derive(Debug, Copy, Clone)]
@@ -115,7 +76,10 @@ pub struct MetaPageData {
     /// now we use advisory locks
     _dead_space_4: [pg_sys::BlockNumber; 2],
 
-    /// pg_search version that created this index. All zeros = created before stamping was added.
+    /// pg_search version that created this index.
+    /// PageInit zeroes the page at creation, so bytes past the fields that exist at creation time
+    /// are zero forever. So we can reliably say:
+    /// All zeros = created before stamping was added.
     created_by_version_major: u16,
     created_by_version_minor: u16,
     created_by_version_patch: u16,
@@ -283,7 +247,6 @@ impl MetaPage {
 
     /// The pg_search version that created this index. Returns `None` for indices created
     /// before version stamping was added (the on-disk fields read as zero).
-    /// The first version with version-stamping is v0.24.0
     #[allow(dead_code)]
     pub fn created_by_version(&self) -> Option<Version> {
         let major = self.data.created_by_version_major;
@@ -426,14 +389,6 @@ unsafe fn bgmerger_state(
 mod tests {
     use super::*;
     use pgrx::prelude::*;
-
-    #[test]
-    fn version_ordering_is_lexicographic_on_components() {
-        assert!(Version::new(0, 18, 0) < Version::new(0, 18, 1));
-        assert!(Version::new(0, 18, 9) < Version::new(0, 19, 0));
-        assert!(Version::new(0, 99, 99) < Version::new(1, 0, 0));
-        assert_eq!(Version::new(1, 2, 3), Version::new(1, 2, 3));
-    }
 
     #[pg_test]
     fn created_by_version_is_stamped_at_index_build() {
